[coc] Add 'coc-python', 'coc-vimlsp', 'coc-json' to auto setup
[autosetup] Mkdir '~/.config' before make neovim soft link
[autoformat] Add '<silent>' to mapping
[script] Refine 'cppcheck' 'buildifier' 'ccls' script

+++ setup/install_cppcheck.sh
@@ -13,15 +13,21 @@ set -v
 VERSION="1.89"
 TARFILE=$VERSION.tar.gz
 DIR=cppcheck-$VERSION
-sudo apt-get update
-sudo apt-get install -y cmake make php7.0-xml
 
-wget https://github.com/danmar/cppcheck/archive/$TARFILE
+if [ ! -d $DIR ]; then
+    sudo apt-get update
+    sudo apt-get install -y \
+        cmake \
+        make \
+        php7.0-xml
 
-tar -xvf $TARFILE
-rm $TARFILE
+    wget https://github.com/danmar/cppcheck/archive/$TARFILE
+    tar -xvf $TARFILE
+    rm $TARFILE
+fi
+
 pushd $DIR > /dev/null
-mkdir build
+mkdir -p build
 pushd build > /dev/null
 cmake ..
 make -j
